# Jaeger Exporter Performance