--- a/src/OpenTelemetry.Exporter.Jaeger/JaegerExporter.cs
+++ b/src/OpenTelemetry.Exporter.Jaeger/JaegerExporter.cs
@@ -17,52 +17,78 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
+using System.Threading;
 using OpenTelemetry.Exporter.Jaeger.Implementation;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
+using Thrift.Protocol;
+using Thrift.Transport;
 
 namespace OpenTelemetry.Exporter.Jaeger
 {
     public class JaegerExporter : ActivityExporter
     {
+        private readonly int maxPacketSize;
+        private readonly TProtocolFactory protocolFactory;
+        private readonly TTransport clientTransport;
+        private readonly JaegerThriftClient thriftClient;
+        private readonly InMemoryTransport memoryTransport;
+        private readonly TProtocol memoryProtocol;
+        private Dictionary<string, Process> processCache;
         private bool libraryResourceApplied;
+        private int batchByteSize;
         private bool disposedValue; // To detect redundant dispose calls
 
-        public JaegerExporter(JaegerExporterOptions options)
+        internal JaegerExporter(JaegerExporterOptions options, TTransport clientTransport = null)
         {
-            this.JaegerAgentUdpBatcher = new JaegerUdpBatcher(options);
+            if (options is null)
+            {
+                throw new ArgumentNullException(nameof(options));
+            }
+
+            this.maxPacketSize = (!options.MaxPacketSize.HasValue || options.MaxPacketSize <= 0) ? JaegerExporterOptions.DefaultMaxPacketSize : options.MaxPacketSize.Value;
+            this.protocolFactory = new TCompactProtocol.Factory();
+            this.clientTransport = clientTransport ?? new JaegerThriftClientTransport(options.AgentHost, options.AgentPort);
+            this.thriftClient = new JaegerThriftClient(this.protocolFactory.GetProtocol(this.clientTransport));
+            this.memoryTransport = new InMemoryTransport(16000);
+            this.memoryProtocol = this.protocolFactory.GetProtocol(this.memoryTransport);
+
+            this.Process = new Process(options.ServiceName, options.ProcessTags);
         }
 
-        internal JaegerExporter(JaegerUdpBatcher jaegerUdpBatcher)
-        {
-            this.JaegerAgentUdpBatcher = jaegerUdpBatcher;
-        }
+        public Process Process { get; internal set; }
 
-        internal JaegerUdpBatcher JaegerAgentUdpBatcher { get; }
+        internal Dictionary<string, Batch> CurrentBatches { get; } = new Dictionary<string, Batch>();
 
         /// <inheritdoc/>
         public override ExportResult Export(in Batch<Activity> activityBatch)
         {
-            var activities = new List<Activity>();
-            foreach (var activity in activityBatch)
+            try
             {
-                activities.Add(activity);
-            }
+                foreach (var activity in activityBatch)
+                {
+                    if (!this.libraryResourceApplied)
+                    {
+                        var libraryResource = activity.GetResource();
 
-            if (!this.libraryResourceApplied && activities.Any())
+                        this.ApplyLibraryResource(libraryResource ?? Resource.Empty);
+
+                        this.libraryResourceApplied = true;
+                    }
+
+                    this.AppendSpan(activity.ToJaegerSpan());
+                }
+
+                this.SendCurrentBatches();
+
+                return ExportResult.Success;
+            }
+            catch (Exception ex)
             {
-                var libraryResource = activities.First().GetResource();
+                JaegerExporterEventSource.Log.FailedExport(ex);
 
-                this.ApplyLibraryResource(libraryResource ?? Resource.Empty);
-
-                this.libraryResourceApplied = true;
+                return ExportResult.Failure;
             }
-
-            _ = this.JaegerAgentUdpBatcher.AppendBatchAsync(activities, default).GetAwaiter().GetResult();
-
-            // TODO jaeger status to ExportResult
-            return ExportResult.Success;
         }
 
         internal void ApplyLibraryResource(Resource libraryResource)
@@ -72,7 +98,7 @@ namespace OpenTelemetry.Exporter.Jaeger
                 throw new ArgumentNullException(nameof(libraryResource));
             }
 
-            var process = this.JaegerAgentUdpBatcher.Process;
+            var process = this.Process;
 
             string serviceName = null;
             string serviceNamespace = null;
@@ -118,25 +144,119 @@ namespace OpenTelemetry.Exporter.Jaeger
         }
 
         /// <inheritdoc/>
-        protected override void OnShutdown(int timeoutMilliseconds)
-        {
-            _ = this.JaegerAgentUdpBatcher.FlushAsync(default).GetAwaiter().GetResult();
-        }
-
-        /// <inheritdoc/>
         protected override void Dispose(bool disposing)
         {
             if (!this.disposedValue)
             {
                 if (disposing)
                 {
-                    this.JaegerAgentUdpBatcher.Dispose();
+                    this.thriftClient.Dispose();
+                    this.clientTransport.Dispose();
+                    this.memoryTransport.Dispose();
+                    this.memoryProtocol.Dispose();
                 }
 
                 this.disposedValue = true;
             }
 
             base.Dispose(disposing);
+        }
+
+        private void AppendSpan(JaegerSpan jaegerSpan)
+        {
+            if (this.processCache == null)
+            {
+                this.Process.Message = this.BuildThriftMessage(this.Process).ToArray();
+                this.processCache = new Dictionary<string, Process>
+                {
+                    [this.Process.ServiceName] = this.Process,
+                };
+            }
+
+            var spanServiceName = jaegerSpan.PeerServiceName ?? this.Process.ServiceName;
+
+            if (!this.processCache.TryGetValue(spanServiceName, out var spanProcess))
+            {
+                spanProcess = new Process(spanServiceName, this.Process.Tags);
+                spanProcess.Message = this.BuildThriftMessage(spanProcess).ToArray();
+                this.processCache.Add(spanServiceName, spanProcess);
+            }
+
+            var spanMessage = this.BuildThriftMessage(jaegerSpan);
+
+            jaegerSpan.Return();
+
+            var spanTotalBytesNeeded = spanMessage.Count;
+
+            if (!this.CurrentBatches.TryGetValue(spanServiceName, out var spanBatch))
+            {
+                spanBatch = new Batch(spanProcess);
+                this.CurrentBatches.Add(spanServiceName, spanBatch);
+
+                spanTotalBytesNeeded += spanProcess.Message.Length;
+            }
+
+            if (this.batchByteSize + spanTotalBytesNeeded >= this.maxPacketSize)
+            {
+                this.SendCurrentBatches();
+
+                // Flushing effectively erases the spanBatch we were working on, so we have to rebuild it.
+                spanBatch.Clear();
+                spanTotalBytesNeeded = spanMessage.Count + spanProcess.Message.Length;
+                this.CurrentBatches.Add(spanServiceName, spanBatch);
+            }
+
+            spanBatch.Add(spanMessage);
+            this.batchByteSize += spanTotalBytesNeeded;
+        }
+
+        private void SendCurrentBatches()
+        {
+            try
+            {
+                foreach (var batch in this.CurrentBatches)
+                {
+                    var task = this.thriftClient.WriteBatchAsync(batch.Value, CancellationToken.None);
+#if DEBUG
+                    if (task.Status != TaskStatus.RanToCompletion)
+                    {
+                        throw new InvalidOperationException();
+                    }
+#endif
+                    batch.Value.Return();
+                }
+            }
+            finally
+            {
+                this.CurrentBatches.Clear();
+                this.batchByteSize = 0;
+                this.memoryTransport.Reset();
+            }
+        }
+
+        private BufferWriterMemory BuildThriftMessage(Process process)
+        {
+            var task = process.WriteAsync(this.memoryProtocol, CancellationToken.None);
+#if DEBUG
+            if (task.Status != TaskStatus.RanToCompletion)
+            {
+                throw new InvalidOperationException();
+            }
+#endif
+            return this.memoryTransport.ToBuffer();
+        }
+
+        // Prevents boxing of JaegerSpan struct.
+        private BufferWriterMemory BuildThriftMessage(in JaegerSpan jaegerSpan)
+        {
+            var task = jaegerSpan.WriteAsync(this.memoryProtocol, CancellationToken.None);
+#if DEBUG
+            if (task.Status != TaskStatus.RanToCompletion)
+            {
+                throw new InvalidOperationException();
+            }
+#endif
+            return this.memoryTransport.ToBuffer();
         }
     }
 }