… src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerUdpBatcher.cs
… src/OpenTelemetry/AssemblyInfo.cs
… src/OpenTelemetry/Internal/PooledList.cs
… test/OpenTelemetry.Exporter.Jaeger.Tests/Implementation/JaegerUdpBatcherTests.cs
… test/OpenTelemetry.Exporter.Jaeger.Tests/JaegerExporterTests.cs