## Changes

* Removed the extra batching going on in Jaeger Exporter. It isn't really needed now that BatchExportActivityProcessor is doing the job. When it is given a batch to export it will build one or more UDP packets depending on if the data <= the max packet size option or not.
* Fixed up the allocations so memory usage is ~flat again. The idea is for the exporter to allocate all the memory it needs for a full batch and then use that memory over and over again.

Performance before:
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |    Error |   StdDev |     Gen 0 |     Gen 1 |     Gen 2 |   Allocated |
|------------------------ |---------------- |-------------- |------------:|---------:|---------:|----------:|----------:|----------:|------------:|
| JaegerExporter_Batching |               1 |         10000 |    120.9 ms |  1.78 ms |  1.58 ms |         - |         - |         - |   512.27 KB |
| JaegerExporter_Batching |              10 |         10000 |  1,182.4 ms | 12.02 ms | 11.25 ms |         - |         - |         - |  2901.73 KB |
| JaegerExporter_Batching |             100 |         10000 | 11,772.2 ms | 99.48 ms | 93.05 ms | 4000.0000 | 4000.0000 | 4000.0000 | 26792.54 KB |

Performance after:
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------:|------:|------:|----------:|
| JaegerExporter_Batching |               1 |         10000 |    119.2 ms |   2.38 ms |   2.64 ms |     - |     - |     - | 236.35 KB |
| JaegerExporter_Batching |              10 |         10000 |  1,160.6 ms |  14.53 ms |  12.13 ms |     - |     - |     - | 238.03 KB |
| JaegerExporter_Batching |             100 |         10000 | 11,700.6 ms | 142.61 ms | 133.40 ms |     - |     - |     - | 240.22 KB |

* [X] `CHANGELOG.md` updated for non-trivial changes
* [X] Unit tests cleanup.