@CodeBlanch 2020-09-13
@reyang I've been thinking about your comment. It's definitely an interesting idea. One thing to note, it's not just the service name we send in that batch header. Process can also have tags. Here's where the message is constructed:

https://github.com/open-telemetry/opentelemetry-dotnet/blob/768eaa3c968197cd06283ef3d24a36d687ce3a84/src/OpenTelemetry.Exporter.Jaeger/Process.cs#L70-L118

There's probably a sweet spot where the process data is small enough we should duplicate it on every span? We cache the bytes for those messages so it would be an easy check to add, just not sure how to figure out the size for the sweet spot.

I just made a version that sends a Jaeger Batch (process + span) for every Activity:

Before:
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |      Median | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------------:|------:|------:|------:|----------:|
| JaegerExporter_Batching |               1 |         10000 |    121.4 ms |   2.33 ms |   2.29 ms |    122.1 ms |     - |     - |     - | 236.35 KB |
| JaegerExporter_Batching |              10 |         10000 |  1,207.3 ms |  24.08 ms |  65.52 ms |  1,179.5 ms |     - |     - |     - | 236.61 KB |
| JaegerExporter_Batching |             100 |         10000 | 11,968.4 ms | 230.42 ms | 256.12 ms | 11,997.0 ms |     - |     - |     - | 252.54 KB |

After:
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |   StdDev |      Median | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|---------:|------------:|------:|------:|------:|----------:|
| JaegerExporter_Batching |               1 |         10000 |    129.1 ms |   2.57 ms |  5.07 ms |    126.6 ms |     - |     - |     - |  17.27 KB |
| JaegerExporter_Batching |              10 |         10000 |  1,243.3 ms |   9.24 ms |  7.72 ms |  1,244.6 ms |     - |     - |     - |  17.36 KB |
| JaegerExporter_Batching |             100 |         10000 | 12,502.6 ms | 120.67 ms | 94.21 ms | 12,508.5 ms |     - |     - |     - |  17.36 KB |

Much less memory used up in the process, but more bytes written to the socket (not tracked in the benchmark). A bit slower too, which was unexpected. Not sure if that is statistically significant. Could be the result of copying many more small buffers to the socket?

Where do you think we should go from here?