@reyang 2020-09-14
> Much less memory used up in the process, but more bytes written to the socket (not tracked in the benchmark). A bit slower too, which was unexpected. Not sure if that is statistically significant. Could be the result of copying many more small buffers to the socket?

This could be tested by replacing the actual socket syscall with a "drop-on-the-floor" dummy function?

> Where do you think we should go from here?

The current perf results doesn't give a strong enough reason to make the change. Where is the perf test code (that has the "process + span" change)? - I wish to take a look.