@CodeBlanch 2020-09-15
@reyang 

> This could be tested by replacing the actual socket syscall with a "drop-on-the-floor" dummy function?

Sorry, my statement was misleading. The benchmark isn't actually writing to a socket, it writes to a "black hole" transport which just drops the bytes. I should have said: "Could be the result of copying many more small buffers to the stream?" That is more accurate with what is actually going on, I think.

> Where is the perf test code

Check out this branch: https://github.com/CodeBlanch/opentelemetry-dotnet/tree/jaeger-process-batch

Luckily I stashed the changes when I was doing this.